+++ src/main/java/fi/fmi/avi/model/GenericAviationWeatherMessage.java
@@ -1,9 +1,10 @@
 package fi.fmi.avi.model;
 
+import java.util.Map;
 import java.util.Optional;
 
 public interface GenericAviationWeatherMessage extends AviationWeatherMessage {
-    enum Format { TAC, IWXXM }
+    Optional<String> getXMLNamespace();
 
     String getOriginalMessage();
 
@@ -13,8 +14,17 @@ public interface GenericAviationWeatherMessage extends AviationWeatherMessage {
 
     Optional<PartialOrCompleteTimePeriod> getValidityTime();
 
-    Optional<Aerodrome> getTargetAerodrome();
+    Map<LocationIndicatorType, String> getLocationIndicators();
 
+    enum Format {
+        TAC, IWXXM
+    }
 
-
+    enum LocationIndicatorType {
+        AERODROME, //
+        ISSUING_CENTRE, //
+        ISSUING_AIR_TRAFFIC_SERVICES_UNIT, //
+        ISSUING_AIR_TRAFFIC_SERVICES_REGION, //
+        ORIGINATING_METEOROLOGICAL_WATCH_OFFICE
+    }
 }